🚚 rename: 폴더 이름 단수형으로 수정
