+++ __tests__/ApplicationTest.js
@@ -1,4 +1,4 @@
-import App from "../src/App.js";
+import App from "../src/app/App.js";
 import { MissionUtils } from "@woowacourse/mission-utils";
 
 const mockQuestions = (inputs) => {
